support jsr310 dates in feign client
See #2874
