+++ modules/swagger-codegen/src/main/resources/Java/libraries/feign/ApiClient.mustache
@@ -9,7 +9,12 @@ import org.apache.oltu.oauth2.client.request.OAuthClientRequest.TokenRequestBuil
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
+{{^java8}}
 import com.fasterxml.jackson.datatype.joda.JodaModule;
+{{/java8}}
+{{#java8}}
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+{{/java8}}
 
 import feign.Feign;
 import feign.RequestInterceptor;
@@ -131,7 +136,12 @@ public class ApiClient {
     objectMapper.enable(DeserializationFeature.READ_ENUMS_USING_TO_STRING);
     objectMapper.disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
     objectMapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    {{^java8}}
     objectMapper.registerModule(new JodaModule());
+    {{/java8}}
+    {{#java8}}
+    objectMapper.registerModule(new JavaTimeModule());
+    {{/java8}}
     return objectMapper;
   }
