+++ samples/client/petstore/java/feign/.swagger-codegen-ignore
@@ -14,7 +14,7 @@
 
 # You can recursively match patterns against a directory, file or extension with a double asterisk (**):
 #foo/**/qux
-# Thsi matches foo/bar/qux, foo/baz/qux, and foo/bar/baz/qux
+# This matches foo/bar/qux, foo/baz/qux, and foo/bar/baz/qux
 
 # You can also negate patterns with an exclamation (!).
 # For example, you can ignore all files in a docs folder with the file extension .md:
