+++ samples/client/petstore/java/feign/build.sbt
@@ -13,11 +13,10 @@ lazy val root = (project in file(".")).
       "com.netflix.feign" % "feign-core" % "8.16.0" % "compile",
       "com.netflix.feign" % "feign-jackson" % "8.16.0" % "compile",
       "com.netflix.feign" % "feign-slf4j" % "8.16.0" % "compile",
-      "com.fasterxml.jackson.core" % "jackson-core" % "2.7.0" % "compile",
-      "com.fasterxml.jackson.core" % "jackson-annotations" % "2.7.0" % "compile",
-      "com.fasterxml.jackson.core" % "jackson-databind" % "2.7.0" % "compile",
-      "com.fasterxml.jackson.datatype" % "jackson-datatype-joda" % "2.1.5" % "compile",
-      "joda-time" % "joda-time" % "2.9.3" % "compile",
+      "com.fasterxml.jackson.core" % "jackson-core" % "2.7.5" % "compile",
+      "com.fasterxml.jackson.core" % "jackson-annotations" % "2.7.5" % "compile",
+      "com.fasterxml.jackson.core" % "jackson-databind" % "2.7.5" % "compile",
+      "com.fasterxml.jackson.datatype" % "jackson-datatype-joda" % "2.7.5" % "compile",
       "org.apache.oltu.oauth2" % "org.apache.oltu.oauth2.client" % "1.0.1" % "compile",
       "com.brsanthu" % "migbase64" % "2.2" % "compile",
       "junit" % "junit" % "4.12" % "test",
