+++ samples/client/petstore/java/feign/src/test/java/io/swagger/client/api/PetApiTest.java
@@ -1,137 +1,200 @@
 package io.swagger.client.api;
 
+import io.swagger.TestUtils;
+
 import io.swagger.client.ApiClient;
-import io.swagger.client.model.Pet;
-import io.swagger.client.model.ModelApiResponse;
+import io.swagger.client.api.*;
+import io.swagger.client.model.*;
+
+import java.io.BufferedWriter;
 import java.io.File;
-import org.junit.Before;
-import org.junit.Test;
-
+import java.io.FileWriter;
 import java.util.ArrayList;
-import java.util.HashMap;
+import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 
-/**
- * API tests for PetApi
- */
+import org.junit.*;
+import static org.junit.Assert.*;
+
 public class PetApiTest {
-
-    private PetApi api;
+    ApiClient apiClient;
+    PetApi api;
 
     @Before
     public void setup() {
-        api = new ApiClient().buildClient(PetApi.class);
+        apiClient = new ApiClient();
+        api = apiClient.buildClient(PetApi.class);
     }
 
-    
-    /**
-     * Add a new pet to the store
-     *
-     * 
-     */
     @Test
-    public void addPetTest() {
-        Pet body = null;
-        // api.addPet(body);
+    public void testApiClient() {
+        // the default api client is used
+        assertEquals("http://petstore.swagger.io/v2", apiClient.getBasePath());
 
-        // TODO: test validations
+        ApiClient newClient = new ApiClient();
+        newClient.setBasePath("http://example.com");
+
+        assertEquals("http://example.com", newClient.getBasePath());
     }
-    
-    /**
-     * Deletes a pet
-     *
-     * 
-     */
+
     @Test
-    public void deletePetTest() {
-        Long petId = null;
-        String apiKey = null;
-        // api.deletePet(petId, apiKey);
+    public void testCreateAndGetPet() throws Exception {
+        Pet pet = createRandomPet();
+        api.addPet(pet);
 
-        // TODO: test validations
+        Pet fetched = api.getPetById(pet.getId());
+        assertNotNull(fetched);
+        assertEquals(pet.getId(), fetched.getId());
+        assertNotNull(fetched.getCategory());
+        assertEquals(fetched.getCategory().getName(), pet.getCategory().getName());
     }
-    
-    /**
-     * Finds Pets by status
-     *
-     * Multiple status values can be provided with comma separated strings
-     */
+
     @Test
-    public void findPetsByStatusTest() {
-        List<String> status = null;
-        // List<Pet> response = api.findPetsByStatus(status);
+    public void testUpdatePet() throws Exception {
+        Pet pet = createRandomPet();
+        pet.setName("programmer");
 
-        // TODO: test validations
+        api.updatePet(pet);
+
+        Pet fetched = api.getPetById(pet.getId());
+        assertNotNull(fetched);
+        assertEquals(pet.getId(), fetched.getId());
+        assertNotNull(fetched.getCategory());
+        assertEquals(fetched.getCategory().getName(), pet.getCategory().getName());
     }
-    
-    /**
-     * Finds Pets by tags
-     *
-     * Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.
-     */
+
     @Test
-    public void findPetsByTagsTest() {
-        List<String> tags = null;
-        // List<Pet> response = api.findPetsByTags(tags);
+    public void testFindPetsByStatus() throws Exception {
+        Pet pet = createRandomPet();
+        pet.setName("programmer");
+        pet.setStatus(Pet.StatusEnum.AVAILABLE);
 
-        // TODO: test validations
+        api.updatePet(pet);
+
+        List<Pet> pets = api.findPetsByStatus(Arrays.asList(new String[]{"available"}));
+        assertNotNull(pets);
+
+        boolean found = false;
+        for (Pet fetched : pets) {
+            if (fetched.getId().equals(pet.getId())) {
+                found = true;
+                break;
+            }
+        }
+
+        assertTrue(found);
     }
-    
-    /**
-     * Find pet by ID
-     *
-     * Returns a single pet
-     */
+
     @Test
-    public void getPetByIdTest() {
-        Long petId = null;
-        // Pet response = api.getPetById(petId);
+    public void testFindPetsByTags() throws Exception {
+        Pet pet = createRandomPet();
+        pet.setName("monster");
+        pet.setStatus(Pet.StatusEnum.AVAILABLE);
 
-        // TODO: test validations
+        List<Tag> tags = new ArrayList<Tag>();
+        Tag tag1 = new Tag();
+        tag1.setName("friendly");
+        tags.add(tag1);
+        pet.setTags(tags);
+
+        api.updatePet(pet);
+
+        List<Pet> pets = api.findPetsByTags(Arrays.asList(new String[]{"friendly"}));
+        assertNotNull(pets);
+
+        boolean found = false;
+        for (Pet fetched : pets) {
+            if (fetched.getId().equals(pet.getId())) {
+                found = true;
+                break;
+            }
+        }
+        assertTrue(found);
     }
-    
-    /**
-     * Update an existing pet
-     *
-     * 
-     */
+
     @Test
-    public void updatePetTest() {
-        Pet body = null;
-        // api.updatePet(body);
+    public void testUpdatePetWithForm() throws Exception {
+        Pet pet = createRandomPet();
+        pet.setName("frank");
+        api.addPet(pet);
 
-        // TODO: test validations
+        Pet fetched = api.getPetById(pet.getId());
+
+        api.updatePetWithForm(fetched.getId(), "furt", null);
+        Pet updated = api.getPetById(fetched.getId());
+
+        assertEquals(updated.getName(), "furt");
     }
-    
-    /**
-     * Updates a pet in the store with form data
-     *
-     * 
-     */
+
     @Test
-    public void updatePetWithFormTest() {
-        Long petId = null;
-        String name = null;
-        String status = null;
-        // api.updatePetWithForm(petId, name, status);
+    public void testDeletePet() throws Exception {
+        Pet pet = createRandomPet();
+        api.addPet(pet);
 
-        // TODO: test validations
+        Pet fetched = api.getPetById(pet.getId());
+        api.deletePet(fetched.getId(), null);
+
+        try {
+            fetched = api.getPetById(fetched.getId());
+            fail("expected an error");
+        } catch (Exception e) {
+//            assertEquals(404, e.getCode());
+        }
     }
-    
-    /**
-     * uploads an image
-     *
-     * 
-     */
+
     @Test
-    public void uploadFileTest() {
-        Long petId = null;
-        String additionalMetadata = null;
-        File file = null;
-        // ModelApiResponse response = api.uploadFile(petId, additionalMetadata, file);
+    public void testUploadFile() throws Exception {
+        Pet pet = createRandomPet();
+        api.addPet(pet);
 
-        // TODO: test validations
+        File file = new File("hello.txt");
+        BufferedWriter writer = new BufferedWriter(new FileWriter(file));
+        writer.write("Hello world!");
+        writer.close();
+
+        api.uploadFile(pet.getId(), "a test file", new File(file.getAbsolutePath()));
     }
-    
+
+    @Test
+    public void testEqualsAndHashCode() {
+        Pet pet1 = new Pet();
+        Pet pet2 = new Pet();
+        assertTrue(pet1.equals(pet2));
+        assertTrue(pet2.equals(pet1));
+        assertTrue(pet1.hashCode() == pet2.hashCode());
+        assertTrue(pet1.equals(pet1));
+        assertTrue(pet1.hashCode() == pet1.hashCode());
+
+        pet2.setName("really-happy");
+        pet2.setPhotoUrls(Arrays.asList(new String[]{"http://foo.bar.com/1", "http://foo.bar.com/2"}));
+        assertFalse(pet1.equals(pet2));
+        assertFalse(pet2.equals(pet1));
+        assertFalse(pet1.hashCode() == (pet2.hashCode()));
+        assertTrue(pet2.equals(pet2));
+        assertTrue(pet2.hashCode() == pet2.hashCode());
+
+        pet1.setName("really-happy");
+        pet1.setPhotoUrls(Arrays.asList(new String[]{"http://foo.bar.com/1", "http://foo.bar.com/2"}));
+        assertTrue(pet1.equals(pet2));
+        assertTrue(pet2.equals(pet1));
+        assertTrue(pet1.hashCode() == pet2.hashCode());
+        assertTrue(pet1.equals(pet1));
+        assertTrue(pet1.hashCode() == pet1.hashCode());
+    }
+
+    private Pet createRandomPet() {
+        Pet pet = new Pet();
+        pet.setId(TestUtils.nextId());
+        pet.setName("gorilla");
+
+        Category category = new Category();
+        category.setName("really-happy");
+
+        pet.setCategory(category);
+        pet.setStatus(Pet.StatusEnum.AVAILABLE);
+        List<String> photos = Arrays.asList(new String[]{"http://foo.bar.com/1", "http://foo.bar.com/2"});
+        pet.setPhotoUrls(photos);
+
+        return pet;
+    }
 }
